requirements.txt
datasets==2.18.0
deepspeed==0.13.1
numpy==1.26.4
peft==0.8.2
Pillow==10.3.0
scikit_learn==1.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.36.2
wandb==0.16.2